==> 1-segmented-sampling-validation/4.phi_Bed_calc_SM.R <==
#reviewing impact of delays in human activity on phi_Bed estimates generated by hourly and segmented method
#apply 1h, 2h, 3h and 6h delays 

require(tidyverse)


#reading in data from ESS systematic review of matched mosquito and human data
df <- read.csv("1-segmented-sampling-validation/data_in/ESS-PNAS-review.csv", header = TRUE)
head(df)
check <- df %>%
  group_by(Ref_name) %>%#
  summarise(PHI_I_check = sum(numeratorPhiI)/sum(denominatorPhiI), 
            PHI_B_check = sum(numerator_phiB)/sum(denominator_PhiB), 
            PHI_I = PHI_I, 
            PHI_B) %>%
  select(Ref_name, PHI_B_check, PHI_B) %>%
  filter(!is.na(PHI_B))

#assign segments####
#evening segment is 19-21h, night is 22-06h and morning is 07-18h. 
df2 <- df %>%
  group_by(Ref_name, Hour) %>%
  mutate(segments = case_when(
    Hour %in% c(19, 20, 21) ~ "19_21", 
    Hour %in% c(22, 23, 24, 1, 2, 3, 4, 5, 6) ~ "22_06", 
    Hour %in% c(7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18) ~ "07_18"))

#read in the prop bed and prop in (human activity) from the systematic review - this is all studies in review containing this data
#uncertainty captured by taking, for each hour, the minimum an maximum prop of people indoors/bed, across all studies
df_bed_review <- read.csv("1-segmented-sampling-validation/data_in/human_in_bed_times_ESS.csv", header = TRUE)
studies <- names(df_bed_review[2:8])
bed_summary <- df_bed_review %>%
  rowwise() %>%
  mutate(prop_Bed_hourly_L = min(c_across(studies)), 
         prop_Bed_hourly_U = max(c_across(studies))) %>%
  select(Hour, prop_Bed_hourly_L, prop_Bed_hourly_U)

df_in_review <- read.csv("1-segmented-sampling-validation/data_in/human_in_times_ESS.csv", header = TRUE)
studies_in <- names(df_in_review[2:23])
in_summary <- df_in_review %>%
  rowwise() %>%
  mutate(prop_In_hourly_L = min(c_across(studies_in)), 
         prop_In_hourly_U = max(c_across(studies_in))) %>%
  select(Hour, prop_In_hourly_L, prop_In_hourly_U)


sleeping_summaries <- left_join(bed_summary, in_summary, by = "Hour")

df2 <- left_join(df2, sleeping_summaries)

#for each study, compute uncertainty in prop_Bed and prop_In
df3_hourly <- df2 %>%
  group_by(Ref_name) %>%
  mutate(
    #uncertainty for hourly phi-B
    numeratorPhiB_L = prop_Bed_hourly_L*Inside_mosq,
    denominatorPhiB_L = ((1 - prop_In_hourly_L)*Outside_mosq) + (prop_In_hourly_L*Inside_mosq),
    numeratorPhiB_U = prop_Bed_hourly_U*Inside_mosq, 
    denominatorPhiB_U = ((1 - prop_In_hourly_U)*Outside_mosq) + (prop_In_hourly_U*Inside_mosq)) %>%
  summarise(PhiB_L  = sum(numeratorPhiB_L)/sum(denominatorPhiB_L), 
            PhiB_U = sum(numeratorPhiB_U)/sum(denominatorPhiB_U), 
            PHI_B = sum(numerator_phiB)/sum(denominator_PhiB))


#delay the human activity data####
#people go inside, go to bed, get out of bed and leave the house 1 hour later
df3_hourly_delay <- df2 %>%
  group_by(Ref_name) %>%
  mutate(
    Hour_1h_earlier = ifelse(Hour > 1, Hour - 1, Hour - 1 + 24), #wrap times around a 24h clock
    Hour_2h_earlier = ifelse(Hour > 2, Hour - 2, Hour - 2 + 24),
    Hour_3h_earlier = ifelse(Hour > 3, Hour - 3, Hour - 3 + 24), 
    Hour_6h_earlier = ifelse(Hour > 6, Hour - 6, Hour - 6 + 24),
  ) %>%
  # Join the original data on these new columns to get the delayed Prop_In and Prop_Bed
  left_join(
    df2 %>%
      select(Hour, Ref_name, Prop_In, prop_In_hourly_L, prop_In_hourly_U, Prop_Bed, prop_Bed_hourly_L, prop_Bed_hourly_U) %>%
      rename(
        Prop_In_1h_delay = Prop_In, Prop_Bed_1h_delay = Prop_Bed, 
        Prop_In_1h_delay_L = prop_In_hourly_L, Prop_Bed_1h_delay_L = prop_Bed_hourly_L, 
        Prop_In_1h_delay_U = prop_In_hourly_U, Prop_Bed_1h_delay_U = prop_Bed_hourly_U
      ),
    by = c("Hour_1h_earlier" = "Hour", "Ref_name")
  ) %>%
  left_join(
    df2 %>%
      select(Hour, Ref_name, Prop_In, prop_In_hourly_L, prop_In_hourly_U, Prop_Bed, prop_Bed_hourly_L, prop_Bed_hourly_U) %>%
      rename(
        Prop_In_2h_delay = Prop_In, Prop_Bed_2h_delay = Prop_Bed,
        Prop_In_2h_delay_L = prop_In_hourly_L, Prop_Bed_2h_delay_L = prop_Bed_hourly_L, 
        Prop_In_2h_delay_U = prop_In_hourly_U, Prop_Bed_2h_delay_U = prop_Bed_hourly_U
      ),
    by = c("Hour_2h_earlier" = "Hour", "Ref_name")
  ) %>%
  left_join(
    df2 %>%
      select(Hour, Ref_name, Prop_In, prop_In_hourly_L, prop_In_hourly_U, Prop_Bed, prop_Bed_hourly_L, prop_Bed_hourly_U) %>%
      rename(
        Prop_In_3h_delay = Prop_In, Prop_Bed_3h_delay = Prop_Bed, 
        Prop_In_3h_delay_L = prop_In_hourly_L, Prop_Bed_3h_delay_L = prop_Bed_hourly_L, 
        Prop_In_3h_delay_U = prop_In_hourly_U, Prop_Bed_3h_delay_U = prop_Bed_hourly_U
      ),
    by = c("Hour_3h_earlier" = "Hour", "Ref_name")
  ) %>%
  left_join(
    df2 %>%
      select(Hour, Ref_name, Prop_In, prop_In_hourly_L, prop_In_hourly_U, Prop_Bed, prop_Bed_hourly_L, prop_Bed_hourly_U) %>%
      rename(
        Prop_In_6h_delay = Prop_In, Prop_Bed_6h_delay = Prop_Bed, 
        Prop_In_6h_delay_L = prop_In_hourly_L, Prop_Bed_6h_delay_L = prop_Bed_hourly_L, 
        Prop_In_6h_delay_U = prop_In_hourly_U, Prop_Bed_6h_delay_U = prop_Bed_hourly_U
      ),
    by = c("Hour_6h_earlier" = "Hour", "Ref_name")
  ) %>%
  
  # Remove the helper columns
  select(-c(Hour_1h_earlier, Hour_2h_earlier, Hour_3h_earlier,  Hour_6h_earlier, 
            Reference_mosquito_data, Reference_human_Data, X, X.1)) %>%
  ungroup()

#inspect delays, can see it works
check_hourly <- df3_hourly_delay %>%
  select(Ref_name, Hour, Prop_In, Prop_In_1h_delay, prop_In_hourly_L, Prop_In_1h_delay_L)

print(check, n = 24)

#add on the segment information 
df3_hourly_delay <- df3_hourly_delay %>%
  mutate(segments = case_when(
    Hour %in% c(19, 20, 21) ~ "19_21", 
    Hour %in% c(22, 23, 24, 1, 2, 3, 4, 5, 6) ~ "22_06", 
    Hour %in% c(7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18) ~ "07_18"))


names(df3_hourly_delay)

plot_data <- df3_hourly_delay %>%
 # filter(Ref_name == "Cooke") %>%
  select(Ref_name, Hour, Inside_mosq, Outside_mosq, Prop_Bed, segments, Prop_Bed_1h_delay, 
         Prop_Bed_2h_delay, Prop_Bed_3h_delay, Prop_Bed_6h_delay)

plot_data_long <- plot_data %>%
  pivot_longer(cols = -c(Hour, Inside_mosq, Outside_mosq, segments, Ref_name), names_to = "delay", values_to = "proportion_bed")




delay_pals <- c('#fb9a99','#a6cee3','#1f78b4','#b2df8a','#33a02c')

plot_data_long <- plot_data_long %>%
  mutate(highlight = case_when(delay == "Prop_Bed" ~ 0, 
                               TRUE ~ 1))
plot_data_long <- plot_data_long %>%
  mutate(combined_var = interaction(as.factor(delay), as.factor(highlight), sep = ", "))


plot_data_long$Hour <- factor(plot_data_long$Hour, 
                              levels = c(19:24, 0:18))  # Custom order


cooke_delays <- plot_data_long %>%
  filter(Ref_name == "Cooke") %>%
  ggplot()+
  aes(x = Hour, y = proportion_bed, col = combined_var, 
                           lty = as.factor(highlight), 
                           group = combined_var)+
  geom_line(size = 1.5)+
  theme_bw()+
  scale_colour_manual(values = delay_pals, name = "Scenario", 
                      labels = c("No delay", "1h delay", "2h delay", 
                                 "3h delay", "6h delay"))+
  guides(lty = "none")+
  ylab("Proportion of people in bed in Cooke study")+
  theme(legend.position = c(0.8, 0.8),
        axis.title = element_text(size = 16),  # Increase font size for axis titles
        axis.text = element_text(size = 16),   # Increase font size for axis labels
        legend.text = element_text(size = 16), # Increase font size for legend text
        legend.title = element_text(size = 16))

#ggsave(cooke_delays, file = "1-segmented-sampling-validation/plots/cooke_delays.png")

plot_data_long %>%
  filter(Ref_name == "SeyoumD") %>%
  select(Inside_mosq, Outside_mosq) %>%
  print(n = 24)

plot_data_long2 <- plot_data_long %>%
  pivot_longer(cols = c(Inside_mosq, Outside_mosq), 
               names_to = "Sampling_location", 
               values_to = "Mosquito_count")

scale_factor <- max(plot_data_long$proportion_bed) / max(plot_data_long$Inside_mosq)
pals <- c('#984ea3','#ff7f00')
delay_pals2 <- c(delay_pals, pals)

time_labels <- c("19:00", "20:00", "21:00", "22:00", "23:00", 
                 "00:00", "01:00", "02:00", "03:00", "04:00", 
                 "05:00", "06:00", "07:00", "08:00", "09:00", 
                 "10:00", "11:00", "12:00", "13:00", "14:00", 
                 "15:00", "16:00", "17:00", "18:00")

SM_figure_data <- ggplot(plot_data_long2, aes(x = Hour, y = proportion_bed, col = combined_var, 
                           lty = as.factor(highlight), 
                           group = combined_var)) +
  geom_line(size = 1.5) +
  
  # Adding mosquito values on a secondary axis
  geom_point(aes(x = Hour, y = Mosquito_count * scale_factor, col = as.factor(Sampling_location)), shape=4, 
             size = 2, stroke = 1.2, alpha = 0.5)+ 
  
  theme_bw() +
  scale_colour_manual(values = delay_pals2, name = "Scenario", 
                      labels = c("No delay", "1h delay", "2h delay", 
                                 "3h delay", "6h delay", "Indoor mosquitoes", "Outdoor mosquitoes")) +
  guides(lty = "none") +
  facet_wrap(vars(Ref_name)) +
  scale_x_discrete(labels = time_labels, name = "Time")+
  ylab("Proportion of people in bed in study") +
  
  scale_y_continuous(
    sec.axis = sec_axis(~ . / scale_factor, name = "Proportion of mosquitoes caught indoors or outdoors") # Transforming back mosquito values
  ) +
  
  theme(legend.position = c(0.8, 0.09),
        axis.title = element_text(size = 12),  # Increase font size for axis titles
        axis.text = element_text(size = 12),   # Increase font size for axis labels
        legend.text = element_text(size = 12), # Increase font size for legend text
        legend.title = element_text(size = 12), 
        axis.text.x = element_text(angle = 90))

ggsave(SM_figure_data, file = "1-segmented-sampling-validation/plots/SM_figure_study_delay.pdf")


#calculate the phi numerator and denominator for each delay scenario####
df3_hourly_delay_phi <- df3_hourly_delay %>%
  group_by(Ref_name) %>%
  mutate(
    
    #uncertainty for 6h delay phi-B
    numeratorPhiB_6h = Prop_Bed_6h_delay*Inside_mosq,
    denominatorPhiB_6h = ((1 - Prop_In_6h_delay)*Outside_mosq) + (Prop_In_6h_delay*Inside_mosq),
    
    numeratorPhiB_6h_L = Prop_Bed_6h_delay_L*Inside_mosq,
    denominatorPhiB_6h_L = ((1 - Prop_In_6h_delay_L)*Outside_mosq) + (Prop_In_6h_delay_L*Inside_mosq),
    
    numeratorPhiB_6h_U = Prop_Bed_6h_delay_U*Inside_mosq, 
    denominatorPhiB_6h_U = ((1 - Prop_In_6h_delay_U)*Outside_mosq) + (Prop_In_6h_delay_U*Inside_mosq),
    
    
    #uncertainty for 3h delay phi-B
    numeratorPhiB_3h = Prop_Bed_3h_delay*Inside_mosq,
    denominatorPhiB_3h = ((1 - Prop_In_3h_delay)*Outside_mosq) + (Prop_In_3h_delay*Inside_mosq),
    
    numeratorPhiB_3h_L = Prop_Bed_3h_delay_L*Inside_mosq,
    denominatorPhiB_3h_L = ((1 - Prop_In_3h_delay_L)*Outside_mosq) + (Prop_In_3h_delay_L*Inside_mosq),
    
    numeratorPhiB_3h_U = Prop_Bed_3h_delay_U*Inside_mosq, 
    denominatorPhiB_3h_U = ((1 - Prop_In_3h_delay_U)*Outside_mosq) + (Prop_In_3h_delay_U*Inside_mosq),
    
    #2h delay 
    
    numeratorPhiB_2h = Prop_Bed_2h_delay*Inside_mosq,
    denominatorPhiB_2h = ((1 - Prop_In_2h_delay)*Outside_mosq) + (Prop_In_2h_delay*Inside_mosq),
    
    numeratorPhiB_2h_L = Prop_Bed_2h_delay_L*Inside_mosq,
    denominatorPhiB_2h_L = ((1 - Prop_In_2h_delay_L)*Outside_mosq) + (Prop_In_2h_delay_L*Inside_mosq),
    
    numeratorPhiB_2h_U = Prop_Bed_2h_delay_U*Inside_mosq, 
    denominatorPhiB_2h_U = ((1 - Prop_In_2h_delay_U)*Outside_mosq) + (Prop_In_2h_delay_U*Inside_mosq), 
    
    #1h delay 
    
    numeratorPhiB_1h = Prop_Bed_1h_delay*Inside_mosq,
    denominatorPhiB_1h = ((1 - Prop_In_1h_delay)*Outside_mosq) + (Prop_In_1h_delay*Inside_mosq),
    
    numeratorPhiB_1h_L = Prop_Bed_1h_delay_L*Inside_mosq,
    denominatorPhiB_1h_L = ((1 - Prop_In_1h_delay_L)*Outside_mosq) + (Prop_In_1h_delay_L*Inside_mosq),
    
    numeratorPhiB_1h_U = Prop_Bed_1h_delay_U*Inside_mosq, 
    denominatorPhiB_1h_U = ((1 - Prop_In_1h_delay_U)*Outside_mosq) + (Prop_In_1h_delay_U*Inside_mosq), 
    
    
  ) %>%
  
  
  summarise(PhiB_3h = sum(numeratorPhiB_3h)/sum(denominatorPhiB_3h), 
            PhiB_3h_L = sum(numeratorPhiB_3h_L)/sum(denominatorPhiB_3h_L), 
            PhiB_3h_U  = sum(numeratorPhiB_3h_U)/sum(denominatorPhiB_3h_U), 
            
            PhiB_2h = sum(numeratorPhiB_2h)/sum(denominatorPhiB_2h), 
            PhiB_2h_L = sum(numeratorPhiB_2h_L)/sum(denominatorPhiB_2h_L), 
            PhiB_2h_U  = sum(numeratorPhiB_2h_U)/sum(denominatorPhiB_2h_U), 
            
            PhiB_1h = sum(numeratorPhiB_1h)/sum(denominatorPhiB_1h), 
            PhiB_1h_L = sum(numeratorPhiB_1h_L)/sum(denominatorPhiB_1h_L), 
            PhiB_1h_U  = sum(numeratorPhiB_1h_U)/sum(denominatorPhiB_1h_U), 
            
            PhiB_6h = sum(numeratorPhiB_6h)/sum(denominatorPhiB_6h), 
            PhiB_6h_L = sum(numeratorPhiB_6h_L)/sum(denominatorPhiB_6h_L), 
            PhiB_6h_U  = sum(numeratorPhiB_6h_U)/sum(denominatorPhiB_6h_U)
            
  )


df4_hourly_phis <- left_join(df3_hourly_delay_phi, df3_hourly, by = "Ref_name")

#all match
df3_hourly$PHI_B == check$PHI_B_check

#segmented method phi calc####

#for segmented method, we sum the proportion of mosquitoes caught indoors and outdoors, in that time segment
#pair the summed proportions with mean proportion of people indoors/bed. uncertainty from human data derived from 20th and 80th quantile of human activity in that segment

df3_seg <- df3_hourly_delay %>%
  group_by(Ref_name, segments) %>%
  summarise(Inside_mosq_seg = sum(Inside_mosq), 
            Outside_mosq_seg = sum(Outside_mosq), 
            tot_mosq_seg = sum(Inside_mosq, Outside_mosq),
            
            #no delay 
            
            prop_In_seg = mean(Prop_In), ##mean
            prop_Bed_seg = mean(Prop_Bed),
            
            prop_Bed_seg_L = quantile(Prop_Bed, 0.2), #for uncertainty for segmented, take the 20 and 80% percentile of human activity for that segment. 
            prop_Bed_seg_U = quantile(Prop_Bed, 0.8),
            prop_In_seg_L = quantile(Prop_In, 0.2), 
            prop_In_seg_U = quantile(Prop_In, 0.8),
            
            
            #6h delay 
            prop_In_6h_seg = mean(Prop_In_6h_delay), ##mean
            prop_Bed_6h_seg = mean(Prop_Bed_6h_delay),
            
            prop_Bed_6h_seg_L = quantile(Prop_Bed_6h_delay, 0.2), #for uncertainty for segmented, take the 20 and 80% percentile for that segment. 
            prop_Bed_6h_seg_U = quantile(Prop_Bed_6h_delay, 0.8),
            
            prop_In_6h_seg_L = quantile(Prop_In_6h_delay, 0.2), 
            prop_In_6h_seg_U = quantile(Prop_In_6h_delay, 0.8),
            
            #3h delay 
            prop_In_3h_seg = mean(Prop_In_3h_delay), ##mean
            prop_Bed_3h_seg = mean(Prop_Bed_3h_delay),
            
            prop_Bed_3h_seg_L = quantile(Prop_Bed_3h_delay, 0.2), #for uncertainty for segmented, take the 20 and 80% percentile for that segment. 
            prop_Bed_3h_seg_U = quantile(Prop_Bed_3h_delay, 0.8),
            
            prop_In_3h_seg_L = quantile(Prop_In_3h_delay, 0.2), 
            prop_In_3h_seg_U = quantile(Prop_In_3h_delay, 0.8),
            
            #2h delay 
            prop_In_2h_seg = mean(Prop_In_2h_delay), ##mean
            prop_Bed_2h_seg = mean(Prop_Bed_2h_delay),
            
            prop_Bed_2h_seg_L = quantile(Prop_Bed_2h_delay, 0.2), #for uncertainty for segmented, take the 20 and 80% percentile for that segment. 
            prop_Bed_2h_seg_U = quantile(Prop_Bed_2h_delay, 0.8),
            
            prop_In_2h_seg_L = quantile(Prop_In_2h_delay, 0.2), 
            prop_In_2h_seg_U = quantile(Prop_In_2h_delay, 0.8),
            
            #1h delay 
            prop_In_1h_seg = mean(Prop_In_1h_delay), ##mean
            prop_Bed_1h_seg = mean(Prop_Bed_1h_delay),
            
            prop_Bed_1h_seg_L = quantile(Prop_Bed_1h_delay, 0.2), #for uncertainty for segmented, take the 20 and 80% percentile for that segment. 
            prop_Bed_1h_seg_U = quantile(Prop_Bed_1h_delay, 0.8),
            
            prop_In_1h_seg_L = quantile(Prop_In_1h_delay, 0.2), 
            prop_In_1h_seg_U = quantile(Prop_In_1h_delay, 0.8),
            
            
  ) %>%
  ungroup()

df3_hourly_data_human <- df3_hourly_delay %>%
  select(Ref_name, Hour, Prop_In_3h_delay, segments)

df3_seg_data_human <- df3_seg %>%
  select(Ref_name,  prop_In_3h_seg, prop_In_3h_seg_L, prop_In_3h_seg_U, segments)

compare_seg_hourly <- left_join(df3_hourly_data_human, df3_seg_data_human, by = c("Ref_name", "segments"))

#quick visual inspection of human activity when segmented or hourly
ggplot(compare_seg_hourly, aes(x = Hour, y = Prop_In_3h_delay))+
  geom_line()+
  geom_line(aes(x = Hour, y = prop_In_3h_seg), col = "blue")+
  facet_wrap(vars(Ref_name))


#then calculate phi numerator and denominator for each segment and study (segmented method)

df3_seg_phi_calc <- df3_seg %>%
  group_by(Ref_name, segments) %>%
  summarise(numerator_phiB_seg = prop_Bed_seg*Inside_mosq_seg, 
            denominator_phiB_seg = ((1-prop_In_seg)*Outside_mosq_seg) + (prop_In_seg*Inside_mosq_seg), 
            
            numerator_phiB_seg_L = prop_Bed_seg_L*Inside_mosq_seg, 
            denominator_phiB_seg_L = ((1-prop_In_seg_L)*Outside_mosq_seg) + (prop_In_seg_L*Inside_mosq_seg), 
            
            numerator_phiB_seg_U = prop_Bed_seg_U*Inside_mosq_seg, 
            denominator_phiB_seg_U = ((1-prop_In_seg_U)*Outside_mosq_seg) + (prop_In_seg_U*Inside_mosq_seg),
            
            #6h delay 
            numerator_phiB_6h_seg = prop_Bed_6h_seg*Inside_mosq_seg, 
            denominator_phiB_6h_seg = ((1-prop_In_6h_seg)*Outside_mosq_seg) + (prop_In_6h_seg*Inside_mosq_seg), 
            
            numerator_phiB_6h_seg_L = prop_Bed_6h_seg_L*Inside_mosq_seg, 
            denominator_phiB_6h_seg_L = ((1-prop_In_6h_seg_L)*Outside_mosq_seg) + (prop_In_6h_seg_L*Inside_mosq_seg), 
            
            numerator_phiB_6h_seg_U = prop_Bed_6h_seg_U*Inside_mosq_seg, 
            denominator_phiB_6h_seg_U = ((1-prop_In_6h_seg_U)*Outside_mosq_seg) + (prop_In_6h_seg_U*Inside_mosq_seg),
            
            
            #3h delay 
            numerator_phiB_3h_seg = prop_Bed_3h_seg*Inside_mosq_seg, 
            denominator_phiB_3h_seg = ((1-prop_In_3h_seg)*Outside_mosq_seg) + (prop_In_3h_seg*Inside_mosq_seg), 
            
            numerator_phiB_3h_seg_L = prop_Bed_3h_seg_L*Inside_mosq_seg, 
            denominator_phiB_3h_seg_L = ((1-prop_In_3h_seg_L)*Outside_mosq_seg) + (prop_In_3h_seg_L*Inside_mosq_seg), 
            
            numerator_phiB_3h_seg_U = prop_Bed_3h_seg_U*Inside_mosq_seg, 
            denominator_phiB_3h_seg_U = ((1-prop_In_3h_seg_U)*Outside_mosq_seg) + (prop_In_3h_seg_U*Inside_mosq_seg),
            
            #2h delay 
            numerator_phiB_2h_seg = prop_Bed_2h_seg*Inside_mosq_seg, 
            denominator_phiB_2h_seg = ((1-prop_In_2h_seg)*Outside_mosq_seg) + (prop_In_2h_seg*Inside_mosq_seg), 
            
            numerator_phiB_2h_seg_L = prop_Bed_2h_seg_L*Inside_mosq_seg, 
            denominator_phiB_2h_seg_L = ((1-prop_In_2h_seg_L)*Outside_mosq_seg) + (prop_In_2h_seg_L*Inside_mosq_seg), 
            
            numerator_phiB_2h_seg_U = prop_Bed_2h_seg_U*Inside_mosq_seg, 
            denominator_phiB_2h_seg_U = ((1-prop_In_2h_seg_U)*Outside_mosq_seg) + (prop_In_2h_seg_U*Inside_mosq_seg),
            
            #1h delay 
            numerator_phiB_1h_seg = prop_Bed_1h_seg*Inside_mosq_seg, 
            denominator_phiB_1h_seg = ((1-prop_In_1h_seg)*Outside_mosq_seg) + (prop_In_1h_seg*Inside_mosq_seg), 
            
            numerator_phiB_1h_seg_L = prop_Bed_1h_seg_L*Inside_mosq_seg, 
            denominator_phiB_1h_seg_L = ((1-prop_In_1h_seg_L)*Outside_mosq_seg) + (prop_In_1h_seg_L*Inside_mosq_seg), 
            
            numerator_phiB_1h_seg_U = prop_Bed_1h_seg_U*Inside_mosq_seg, 
            denominator_phiB_1h_seg_U = ((1-prop_In_1h_seg_U)*Outside_mosq_seg) + (prop_In_1h_seg_U*Inside_mosq_seg)
            
  ) %>%
  ungroup()

#calculate phi for each study####

df4_seg_phis <- df3_seg_phi_calc %>%
  group_by(Ref_name) %>%
  summarise(
    PHI_B_seg = sum(numerator_phiB_seg)/sum(denominator_phiB_seg), 
    PHI_B_seg_L  = sum(numerator_phiB_seg_L)/sum(denominator_phiB_seg_L), 
    PHI_B_seg_U = sum(numerator_phiB_seg_U)/sum(denominator_phiB_seg_U),
    
    ##6h delay 
    PHI_B_6h_seg = sum(numerator_phiB_6h_seg)/sum(denominator_phiB_6h_seg), 
    PHI_B_6h_seg_L  = sum(numerator_phiB_6h_seg_L)/sum(denominator_phiB_6h_seg_L), 
    PHI_B_6h_seg_U = sum(numerator_phiB_6h_seg_U)/sum(denominator_phiB_6h_seg_U),
    
    ##3h delay 
    PHI_B_3h_seg = sum(numerator_phiB_3h_seg)/sum(denominator_phiB_3h_seg), 
    PHI_B_3h_seg_L  = sum(numerator_phiB_3h_seg_L)/sum(denominator_phiB_3h_seg_L), 
    PHI_B_3h_seg_U = sum(numerator_phiB_3h_seg_U)/sum(denominator_phiB_3h_seg_U),
    #
    ##2h delay 
    PHI_B_2h_seg = sum(numerator_phiB_2h_seg)/sum(denominator_phiB_2h_seg), 
    PHI_B_2h_seg_L  = sum(numerator_phiB_2h_seg_L)/sum(denominator_phiB_2h_seg_L), 
    PHI_B_2h_seg_U = sum(numerator_phiB_2h_seg_U)/sum(denominator_phiB_2h_seg_U),
    #
    ##1h delay 
    PHI_B_1h_seg = sum(numerator_phiB_1h_seg)/sum(denominator_phiB_1h_seg), 
    PHI_B_1h_seg_L  = sum(numerator_phiB_1h_seg_L)/sum(denominator_phiB_1h_seg_L), 
    PHI_B_1h_seg_U = sum(numerator_phiB_1h_seg_U)/sum(denominator_phiB_1h_seg_U)
    
  )

#processing to ensure lower, median and upper bound

df5_seg_phis <- df4_seg_phis %>%
  group_by(Ref_name) %>%
  summarise(newPHI_B_seg = median(c(PHI_B_seg, PHI_B_seg_L, PHI_B_seg_U)),
            newPHI_B_seg_L = min(c(PHI_B_seg, PHI_B_seg_L, PHI_B_seg_U)), 
            newPHI_B_seg_U = max(c(PHI_B_seg, PHI_B_seg_L, PHI_B_seg_U)), 
            
            newPHI_B_6h_seg = median(c(PHI_B_6h_seg, PHI_B_6h_seg_L, PHI_B_6h_seg_U)), 
            newPHI_B_6h_seg_L = min(c(PHI_B_6h_seg, PHI_B_6h_seg_L, PHI_B_6h_seg_U)),
            newPHI_B_6h_seg_U = max(c(PHI_B_6h_seg, PHI_B_6h_seg_L, PHI_B_6h_seg_U)),
            
            newPHI_B_1h_seg = median(c(PHI_B_1h_seg, PHI_B_1h_seg_L, PHI_B_1h_seg_U)), 
            newPHI_B_1h_seg_L = min(c(PHI_B_1h_seg, PHI_B_1h_seg_L, PHI_B_1h_seg_U)),
            newPHI_B_1h_seg_U = max(c(PHI_B_1h_seg, PHI_B_1h_seg_L, PHI_B_1h_seg_U)),
            
            newPHI_B_2h_seg = median(c(PHI_B_2h_seg, PHI_B_2h_seg_L, PHI_B_2h_seg_U)), 
            newPHI_B_2h_seg_L = min(c(PHI_B_2h_seg, PHI_B_2h_seg_L, PHI_B_2h_seg_U)),
            newPHI_B_2h_seg_U = max(c(PHI_B_2h_seg, PHI_B_2h_seg_L, PHI_B_2h_seg_U)),
            
            newPHI_B_3h_seg = median(c(PHI_B_3h_seg, PHI_B_3h_seg_L, PHI_B_3h_seg_U)), 
            newPHI_B_3h_seg_L = min(c(PHI_B_3h_seg, PHI_B_3h_seg_L, PHI_B_3h_seg_U)),
            newPHI_B_3h_seg_U = max(c(PHI_B_3h_seg, PHI_B_3h_seg_L, PHI_B_3h_seg_U)),
            
  ) 


df4_compare <- left_join(df5_seg_phis, df4_hourly_phis, by = "Ref_name")

colnames(df4_compare) <- gsub("new", "", colnames(df4_compare))

#summary stats to compare segmented and hourly estimates####

#look at residuals from y = x for each (absolute difference)
#also look at relative difference. 
#calculate Rsquared
calc_Rsq <- function(segment, hourly){
  res <- segment - hourly 
  res_sq <- res^2
  SS_res <- sum(res_sq)
  
  SS_tot <- sum((segment-mean(segment))^2)
  R_sq <- 1 - (SS_res/SS_tot)
  
  return(R_sq)
}

#summary stats no delay####
fitted_phiB <- df4_compare$PHI_B
residuals_phiB <- abs(df4_compare$PHI_B_seg-fitted_phiB)
max_res_no_delay <- round(max(abs(residuals_phiB)),3) #0.1010978

r_sq_nodelay <- calc_Rsq(df4_compare$PHI_B_seg, df4_compare$PHI_B)
lab_phi_B_nodelay <- paste("R² =", round(r_sq_nodelay, 2))

rel_diff_no_delay = ((df4_compare$PHI_B_seg - fitted_phiB)/df4_compare$PHI_B_seg)*100

#summary stats 1h delay####
fitted_phiB_1h <- df4_compare$PhiB_1h
residuals_phiB_1h <- abs(df4_compare$PHI_B_1h_seg-fitted_phiB_1h)
max_res_1h_delay <- round(max(abs(residuals_phiB_1h)),3) #0.0246

r_sq_1h_delay <- calc_Rsq(df4_compare$PHI_B_1h_seg, df4_compare$PhiB_1h)
lab_phi_B_1h_delay <- paste("R² =", round(r_sq_1h_delay, 2))

rel_diff_1h_delay = ((df4_compare$PHI_B_1h_seg - fitted_phiB_1h)/df4_compare$PHI_B_1h_seg)*100

#summary stats 2h delay####
fitted_phiB_2h <- df4_compare$PhiB_2h
residuals_phiB_2h <- abs(df4_compare$PHI_B_2h_seg-fitted_phiB_2h)
max_res_2h_delay <- round(max(abs(residuals_phiB_2h)),3) #0.0444

r_sq_2h_delay <- calc_Rsq(df4_compare$PHI_B_2h_seg, df4_compare$PhiB_2h)
lab_phi_B_2h_delay <- paste("R² =", round(r_sq_2h_delay, 2))

rel_diff_2h_delay = ((df4_compare$PHI_B_2h_seg - fitted_phiB_2h)/df4_compare$PHI_B_2h_seg)*100

#summary stats 3h delay####
fitted_phiB_3h <- df4_compare$PhiB_3h
residuals_phiB_3h <- abs(df4_compare$PHI_B_3h_seg-fitted_phiB_3h)
max_res_3h_delay <- round(max(abs(residuals_phiB_3h)),3) #0.083

r_sq_3h_delay <- calc_Rsq(df4_compare$PHI_B_3h_seg, df4_compare$PhiB_3h)
lab_phi_B_3h_delay <- paste("R² =", round(r_sq_3h_delay, 2))
rel_diff_3h_delay = (( df4_compare$PHI_B_3h_seg - fitted_phiB_3h)/df4_compare$PHI_B_3h_seg)*100


#summary stats 6h delay####
fitted_phiB_6h <- df4_compare$PhiB_6h
residuals_phiB_6h <- abs(df4_compare$PHI_B_6h_seg-fitted_phiB_6h)
max_res_6h_delay <- round(max(abs(residuals_phiB_6h)),3) #0.091

r_sq_6h_delay <- calc_Rsq(df4_compare$PHI_B_6h_seg, df4_compare$PhiB_6h)
lab_phi_B_6h_delay <- paste("R² =", round(r_sq_6h_delay, 2))
rel_diff_6h_delay = (( df4_compare$PHI_B_6h_seg - fitted_phiB_6h)/df4_compare$PHI_B_6h_seg)*100


ref <- df4_compare$Ref_name
#relative difference between segmented and hourly method, for each study and delay scenario
errors <- data.frame(ref,
                     rel_diff_no_delay, rel_diff_1h_delay, rel_diff_2h_delay, rel_diff_3h_delay, 
                     rel_diff_6h_delay)


errors_rel_diff <- errors %>%
  pivot_longer(cols = c(rel_diff_no_delay, rel_diff_1h_delay, 
                        rel_diff_2h_delay, rel_diff_3h_delay, 
                        rel_diff_6h_delay),
               names_to = "delay", values_to = "rel_diff")
delay_pals3 <- c('#a6cee3','#1f78b4','#b2df8a','#33a02c', '#fb9a99')

ylab_rel_diff <- expression(atop("Relative difference (%) in" ~ phi[Bed], "(segment vs hourly)"))
col_lab <- "Study from Sherrard-Smith et al. (2019) review"

rel_diff_plot <- ggplot(errors_rel_diff, aes(x = ref, y = rel_diff, fill = as.factor(delay)))+
  geom_bar(stat = "identity", position = position_dodge())+
  ylab(ylab_rel_diff)+
  xlab(col_lab)+
  scale_fill_manual(values = delay_pals3, 
                    labels = c("1h", "2h", "3h","6h", "No delay"), name = "Delay scenario")+
  theme_bw()+
  theme(legend.position = c(0.7, 0.7),
        axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
        text=element_text(size=10), #change font size of all text
        axis.text=element_text(size=12), #change font size of axis text
        axis.title=element_text(size=12), #change font size of axis titles
        plot.title=element_text(size=12), #change font size of plot title
        legend.text=element_text(size=12), #change font size of legend text
        legend.title=element_text(size=12),
        axis.ticks.length = unit(5, "pt"))


xlab <- expression(phi[Bed] ~ "with hourly mosquito data")
ylab <- expression(phi[Bed] ~ "with segmented mosquito data")

#plots show the R2 of hourly vs segmented for each study

no_delay <- ggplot(df4_compare, aes(x = PHI_B, y = PHI_B_seg, fill = as.factor(Ref_name)) 
)+
  geom_point(size = 4,shape = 21, alpha = 0.8)+
  geom_errorbar(aes(ymin = PHI_B_seg_L, ymax = PHI_B_seg_U, col = as.factor(Ref_name)), width = 0.02)+
  geom_errorbarh(aes(xmin = PhiB_L, xmax = PhiB_U, col = as.factor(Ref_name)), height = 0.02)+
  ylim(0,1)+
  xlim(0,1)+
  geom_abline(slope = 1, intercept = 0, col = "red", linetype =  "dashed")+
  theme_bw()+
  theme(legend.position = c(0.2, 0.9),
        text=element_text(size=12), #change font size of all text
        axis.text=element_text(size=12), #change font size of axis text
        axis.title=element_text(size=12), #change font size of axis titles
        plot.title=element_text(size=12), #change font size of plot title
        legend.text=element_text(size=12), #change font size of legend text
        legend.title=element_text(size=12),
        axis.ticks.length = unit(5, "pt"))+
  guides(fill = "none", col = "none")+
  geom_text(aes(x = 0.3, y = 0.8, label = lab_phi_B_nodelay), color = "black")+
  labs(col = col_lab, x = xlab, 
       y = ylab)+
  ggtitle("No delay on human activity")

delay_1h <- ggplot(df4_compare, aes(x = PhiB_1h, y = PHI_B_1h_seg, fill = as.factor(Ref_name)))+
  geom_point(size = 4, shape = 21, alpha = 0.8)+
  geom_errorbarh(aes(xmin = PhiB_1h_L, xmax = PhiB_1h_U, col = as.factor(Ref_name)), height = 0.02)+
  geom_errorbar(aes(ymin = PHI_B_1h_seg_L, ymax = PHI_B_1h_seg_U, col = as.factor(Ref_name)), width = 0.02)+
  ylim(0,1)+
  xlim(0,1)+
  geom_abline(slope = 1, intercept = 0, col = "red", linetype =  "dashed")+
  theme_bw()+
  theme(legend.position = c(0.2, 0.7),
        text=element_text(size=12), #change font size of all text
        axis.text=element_text(size=12), #change font size of axis text
        axis.title=element_text(size=12), #change font size of axis titles
        plot.title=element_text(size=12), #change font size of plot title
        legend.text=element_text(size=12), #change font size of legend text
        legend.title=element_text(size=12),
        axis.ticks.length = unit(5, "pt"))+
  guides(fill = "none", col = "none")+
  geom_text(aes(x = 0.3, y = 0.8, label = lab_phi_B_1h_delay), color = "black")+
  labs(col = col_lab, x = xlab, 
       y = xlab)+
  ggtitle("1h delay on human activity")

delay_2h <- ggplot(df4_compare, aes(x = PhiB_2h, y = PHI_B_2h_seg, fill = as.factor(Ref_name)))+
  geom_point(size = 4, shape = 21, alpha = 0.8)+
  geom_errorbarh(aes(xmin = PhiB_2h_L, xmax = PhiB_2h_U, col = as.factor(Ref_name)), height = 0.02)+
  geom_errorbar(aes(ymin = PHI_B_2h_seg_L, ymax = PHI_B_2h_seg_U, col = as.factor(Ref_name)), width = 0.02)+
  ylim(0,1)+
  xlim(0,1)+
  geom_abline(slope = 1, intercept = 0, col = "red", linetype =  "dashed")+
  theme_bw()+
  theme(legend.position = c(0.2, 0.7),
        text=element_text(size=12), #change font size of all text
        axis.text=element_text(size=12), #change font size of axis text
        axis.title=element_text(size=12), #change font size of axis titles
        plot.title=element_text(size=12), #change font size of plot title
        legend.text=element_text(size=12), #change font size of legend text
        legend.title=element_text(size=12),
        axis.ticks.length = unit(5, "pt"))+
  guides(fill = "none", col = "none")+
  geom_text(aes(x = 0.3, y = 0.8, label = lab_phi_B_2h_delay), color = "black")+
  labs(col = col_lab, x = xlab, 
       y = ylab)+
  ggtitle("2h delay on human activity")


delay_3h <- ggplot(df4_compare, aes(x = PhiB_3h, y = PHI_B_3h_seg, fill = as.factor(Ref_name)))+
  geom_point(size = 4, shape = 21, alpha = 0.8)+
  geom_errorbarh(aes(xmin = PhiB_3h_L, xmax = PhiB_3h_U, col = as.factor(Ref_name)), height = 0.02)+
  geom_errorbar(aes(ymin = PHI_B_3h_seg_L, ymax = PHI_B_3h_seg_U, col = as.factor(Ref_name)), width = 0.02)+
  ylim(0,1)+
  xlim(0,1)+
  geom_abline(slope = 1, intercept = 0, col = "red", linetype =  "dashed")+
  theme_bw()+
  theme(legend.position = c(0.2, 0.7),
        text=element_text(size=12), #change font size of all text
        axis.text=element_text(size=12), #change font size of axis text
        axis.title=element_text(size=12), #change font size of axis titles
        plot.title=element_text(size=12), #change font size of plot title
        legend.text=element_text(size=12), #change font size of legend text
        legend.title=element_text(size=12),
        axis.ticks.length = unit(5, "pt"))+
  guides(fill = "none", col = "none")+
  geom_text(aes(x = 0.3, y = 0.8, label = lab_phi_B_3h_delay), color = "black")+
  labs(col = col_lab, x = xlab, 
       y = ylab)+
  ggtitle("3h delay on human activity")

fill_lab <- "Study from  \n Sherrard-Smith et al. \n (2019) review"


delay_6h <- ggplot(df4_compare, aes(x = PhiB_6h, y = PHI_B_6h_seg, fill = as.factor(Ref_name)))+
  geom_point(size = 4, shape = 21, alpha = 0.8)+
  geom_errorbarh(aes(xmin = PhiB_6h_L, xmax = PhiB_6h_U, col = as.factor(Ref_name)), height = 0.02)+
  geom_errorbar(aes(ymin = PHI_B_6h_seg_L, ymax = PHI_B_6h_seg_U, col = as.factor(Ref_name)), width = 0.02)+
  ylim(0,1)+
  xlim(0,1)+
  geom_abline(slope = 1, intercept = 0, col = "red", linetype =  "dashed")+
  theme_bw()+
  theme(legend.position = c(0.8, 0.4),
        text=element_text(size=10), #change font size of all text
        axis.text=element_text(size=12), #change font size of axis text
        axis.title=element_text(size=12), #change font size of axis titles
        plot.title=element_text(size=12), #change font size of plot title
        legend.text=element_text(size=10), #change font size of legend text
        legend.title=element_text(size=10),
        axis.ticks.length = unit(5, "pt"))+
  guides(col = "none")+
  geom_text(aes(x = 0.3, y = 0.8, label = lab_phi_B_6h_delay), color = "black")+
  labs(fill = fill_lab, x = xlab, 
       y = ylab)+
  ggtitle("6h delay on human activity")



break_segments_plot <- cowplot::plot_grid(no_delay, delay_1h, delay_2h, delay_3h, delay_6h,
                                          rel_diff_plot,
                                          ncol = 3, nrow = 2, 
                                          labels = c("A", "B", "C", "D", "E", "F")) 

ggsave(break_segments_plot, file = "1-segmented-sampling-validation/plots/SM_segment.pdf")


#for presentations (bigger fonts)

no_delay_SM <- ggplot(df4_compare, aes(x = PHI_B, y = PHI_B_seg, fill = as.factor(Ref_name)) 
)+
  geom_point(size = 4,shape = 21, alpha = 0.8)+
  geom_errorbar(aes(ymin = PHI_B_seg_L, ymax = PHI_B_seg_U, col = as.factor(Ref_name)), width = 0.02)+
  geom_errorbarh(aes(xmin = PhiB_L, xmax = PhiB_U, col = as.factor(Ref_name)), height = 0.02)+
  ylim(0,1)+
  xlim(0,1)+
  geom_abline(slope = 1, intercept = 0, col = "red", linetype =  "dashed")+
  theme_bw()+
  theme(legend.position = c(0.2, 0.9),
        text=element_text(size=16), #change font size of all text
        axis.text=element_text(size=16), #change font size of axis text
        axis.title=element_text(size=16), #change font size of axis titles
        plot.title=element_text(size=16), #change font size of plot title
        legend.text=element_text(size=16), #change font size of legend text
        legend.title=element_text(size=16),
        axis.ticks.length = unit(5, "pt"))+
  guides(fill = "none", col = "none")+
  geom_text(aes(x = 0.3, y = 0.8, label = lab_phi_B_nodelay), color = "black")+
  labs(col = col_lab, x = xlab, 
       y = ylab)+
  ggtitle("No delay on human activity")

delay_1h_SM <- ggplot(df4_compare, aes(x = PhiB_1h, y = PHI_B_1h_seg, fill = as.factor(Ref_name)))+
  geom_point(size = 4, shape = 21, alpha = 0.8)+
  geom_errorbarh(aes(xmin = PhiB_1h_L, xmax = PhiB_1h_U, col = as.factor(Ref_name)), height = 0.02)+
  geom_errorbar(aes(ymin = PHI_B_1h_seg_L, ymax = PHI_B_1h_seg_U, col = as.factor(Ref_name)), width = 0.02)+
  ylim(0,1)+
  xlim(0,1)+
  geom_abline(slope = 1, intercept = 0, col = "red", linetype =  "dashed")+
  theme_bw()+
  theme(legend.position = c(0.2, 0.7),
        text=element_text(size=16), #change font size of all text
        axis.text=element_text(size=16), #change font size of axis text
        axis.title=element_text(size=16), #change font size of axis titles
        plot.title=element_text(size=16), #change font size of plot title
        legend.text=element_text(size=16), #change font size of legend text
        legend.title=element_text(size=16),
        axis.ticks.length = unit(5, "pt"))+
  guides(fill = "none", col = "none")+
  geom_text(aes(x = 0.3, y = 0.8, label = lab_phi_B_1h_delay), color = "black")+
  labs(col = col_lab, x = xlab, 
       y = xlab)+
  ggtitle("1h delay on human activity")

delay_2h_SM <- ggplot(df4_compare, aes(x = PhiB_2h, y = PHI_B_2h_seg, fill = as.factor(Ref_name)))+
  geom_point(size = 4, shape = 21, alpha = 0.8)+
  geom_errorbarh(aes(xmin = PhiB_2h_L, xmax = PhiB_2h_U, col = as.factor(Ref_name)), height = 0.02)+
  geom_errorbar(aes(ymin = PHI_B_2h_seg_L, ymax = PHI_B_2h_seg_U, col = as.factor(Ref_name)), width = 0.02)+
  ylim(0,1)+
  xlim(0,1)+
  geom_abline(slope = 1, intercept = 0, col = "red", linetype =  "dashed")+
  theme_bw()+
  theme(legend.position = c(0.2, 0.7),
        text=element_text(size=16), #change font size of all text
        axis.text=element_text(size=16), #change font size of axis text
        axis.title=element_text(size=16), #change font size of axis titles
        plot.title=element_text(size=16), #change font size of plot title
        legend.text=element_text(size=16), #change font size of legend text
        legend.title=element_text(size=16),
        axis.ticks.length = unit(5, "pt"))+
  guides(fill = "none", col = "none")+
  geom_text(aes(x = 0.3, y = 0.8, label = lab_phi_B_2h_delay), color = "black")+
  labs(col = col_lab, x = xlab, 
       y = ylab)+
  ggtitle("2h delay on human activity")


delay_3h_SM <- ggplot(df4_compare, aes(x = PhiB_3h, y = PHI_B_3h_seg, fill = as.factor(Ref_name)))+
  geom_point(size = 4, shape = 21, alpha = 0.8)+
  geom_errorbarh(aes(xmin = PhiB_3h_L, xmax = PhiB_3h_U, col = as.factor(Ref_name)), height = 0.02)+
  geom_errorbar(aes(ymin = PHI_B_3h_seg_L, ymax = PHI_B_3h_seg_U, col = as.factor(Ref_name)), width = 0.02)+
  ylim(0,1)+
  xlim(0,1)+
  geom_abline(slope = 1, intercept = 0, col = "red", linetype =  "dashed")+
  theme_bw()+
  theme(legend.position = c(0.2, 0.7),
        text=element_text(size=16), #change font size of all text
        axis.text=element_text(size=16), #change font size of axis text
        axis.title=element_text(size=16), #change font size of axis titles
        plot.title=element_text(size=16), #change font size of plot title
        legend.text=element_text(size=16), #change font size of legend text
        legend.title=element_text(size=16),
        axis.ticks.length = unit(5, "pt"))+
  guides(fill = "none", col = "none")+
  geom_text(aes(x = 0.3, y = 0.8, label = lab_phi_B_3h_delay), color = "black")+
  labs(col = col_lab, x = xlab, 
       y = ylab)+
  ggtitle("3h delay on human activity")

fill_lab <- "Study"


delay_6h_SM <- ggplot(df4_compare, aes(x = PhiB_6h, y = PHI_B_6h_seg, fill = as.factor(Ref_name)))+
  geom_point(size = 4, shape = 21, alpha = 0.8)+
  geom_errorbarh(aes(xmin = PhiB_6h_L, xmax = PhiB_6h_U, col = as.factor(Ref_name)), height = 0.02)+
  geom_errorbar(aes(ymin = PHI_B_6h_seg_L, ymax = PHI_B_6h_seg_U, col = as.factor(Ref_name)), width = 0.02)+
  ylim(0,1)+
  xlim(0,1)+
  geom_abline(slope = 1, intercept = 0, col = "red", linetype =  "dashed")+
  theme_bw()+
  theme(legend.position = c(0.83, 0.6),
        text=element_text(size=14), #change font size of all text
        axis.text=element_text(size=16), #change font size of axis text
        axis.title=element_text(size=16), #change font size of axis titles
        plot.title=element_text(size=16), #change font size of plot title
        legend.text=element_text(size=14), #change font size of legend text
        legend.title=element_text(size=14),
        axis.ticks.length = unit(5, "pt"))+
  guides(col = "none", fill = "none")+
  geom_text(aes(x = 0.3, y = 0.8, label = lab_phi_B_6h_delay), color = "black")+
  labs(fill = fill_lab, x = xlab, 
       y = ylab)+
  ggtitle("6h delay on human activity")

ylab_rel_diff_SM <- expression(atop("Relative difference (%) in" ~ phi[Bed], "(segment vs hourly)"))


rel_diff_plot_SM <- ggplot(errors_rel_diff, aes(x = ref, y = rel_diff, fill = as.factor(delay)))+
  geom_bar(stat = "identity", position = position_dodge())+
  ylab(ylab_rel_diff_SM)+
  xlab(col_lab)+
  scale_fill_manual(values = delay_pals3, 
                    labels = c("1h", "2h", "3h","6h", "No delay"), name = "Delay scenario")+
  theme_bw()+
  theme(legend.position ="top",
        legend.direction = "horizontal",
        axis.text.x = element_text(angle = 45, hjust=1),
        text=element_text(size=14), #change font size of all text
        axis.text=element_text(size=16), #change font size of axis text
        axis.title=element_text(size=16), #change font size of axis titles
        plot.title=element_text(size=16), #change font size of plot title
        legend.text=element_text(size=16), #change font size of legend text
        legend.title=element_text(size=16),
        axis.ticks.length = unit(5, "pt"))


break_segments_plot_SM_1 <- cowplot::plot_grid(no_delay_SM, delay_6h_SM,
                                             ncol = 2, 
                                             labels = c("C", "D"), hjust = -2.5)

break_segments_plot_SM_2 <- cowplot::plot_grid(rel_diff_plot_SM, break_segments_plot_SM_1,
                                               nrow = 2, 
                                               labels = c("B", ""), hjust = -2.5)


break_segments_plot_SM_3 <- cowplot::plot_grid(cooke_delays, break_segments_plot_SM_2, 
                                               labels = c("A", ""))

#ggsave(break_segments_plot_SM_3, file = "1-segmented-sampling-validation/plots/SM_segment_pres.png")
#ggsave(break_segments_plot_SM_3, file = "1-segmented-sampling-validation/plots/SM_segment_pres.svg", dpi = 300)
